ai-detection-service/requirements.txt
opencv-python>=4.8.0
ultralytics>=8.0.0
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
# Optional: uncomment for face recognition (alert 15-18)
# insightface>=0.7.3
# onnxruntime>=1.16.0
